requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-default-group==1.2.4
coderedcms==4.0.1
coverage==7.6.4
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.3
django-bootstrap5==24.2
django-debug-toolbar==4.4.6
django-eventtools==1.0.3
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-rosetta==0.10.1
django-sass==1.1.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
future==1.0.0
geocoder==1.38.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
icalendar==5.0.13
idna==3.10
iniconfig==2.0.0
jiter==0.7.0
l18n==2021.3
laces==0.1.1
libsass==0.23.0
llm==0.17.1
lxml==5.3.0
mypy==1.13.0
mypy-extensions==1.0.0
openai==1.54.3
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pillow_heif==0.20.0
playwright==1.48.0
pluggy==1.5.0
polib==1.2.0
puremagic==1.28
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
pytesseract==0.3.13
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-ulid==3.0.0
pytz==2024.2
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.3
ruff==0.7.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.37
sqlparse==0.5.1
tabulate==0.9.0
telepath==0.3.1
tqdm==4.67.0
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
urllib3==2.2.3
wagtail==6.3
wagtail-ai==2.1.1
wagtail-cache==2.5.1
wagtail-localize==1.10
wagtail-seo==2.5.0
wagtail_word==1.1.3
wagtailmedia==0.15.2
Willow==1.9.0
